+++ rollup.config.js
@@ -2,6 +2,7 @@ import autoExternal from 'rollup-plugin-auto-external';
 import babel from 'rollup-plugin-babel';
 import resolve from 'rollup-plugin-node-resolve';
 import replace from 'rollup-plugin-replace';
+import json from 'rollup-plugin-json';
 
 import pkg from './package.json';
 
@@ -15,6 +16,10 @@ export default [
             resolve({
                 jsnext: true,
                 extensions,
+            }),
+            json({
+                include: 'src/**',
+                indent: '  ',
             }),
             babel({
                 exclude: 'node_modules/**',