+++ src/config/validate-config.ts
@@ -1,3 +1,6 @@
+import * as ajv from 'ajv';
+import * as schema from './config-schema.json';
+
 export default (foundConfig: any): void => {
     if (!foundConfig) {
         throw new Error('No config found');
@@ -7,7 +10,15 @@ export default (foundConfig: any): void => {
         throw new Error('No config contents found');
     }
 
-    if (!foundConfig.config.root) {
-        throw new Error('No root specified');
+    const validator = new ajv({
+        errorDataPath: 'configuration',
+        allErrors: true,
+    });
+
+    const validate = validator.compile(schema);
+    const valid = validate(foundConfig.config);
+
+    if (!valid && validate.errors) {
+        throw new Error(`Invalid config: ${JSON.stringify(validate.errors, null, 2)}`);
     }
 };